feat: user can set/delete like on card

+++ src/utils/Api.js
@@ -70,18 +70,32 @@ class Api {
     }).then(this._getResponse);
   }
 
-  setLike(cardId) {
-    return fetch(`${this._url}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then(this._getResponse);
-  }
+  // setLike(cardId) {
+  //   return fetch(`${this._url}/cards/${cardId}/likes`, {
+  //     method: "PUT",
+  //     headers: this._headers,
+  //   }).then(this._getResponse);
+  // }
 
-  removeLike(cardId) {
-    return fetch(`${this._url}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then(this._getResponse);
+  // removeLike(cardId) {
+  //   return fetch(`${this._url}/cards/${cardId}/likes`, {
+  //     method: "DELETE",
+  //     headers: this._headers,
+  //   }).then(this._getResponse);
+  // }
+
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._getResponse);
+    } else {
+      return fetch(`${this._url}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._getResponse);
+    }
   }
 }
